Fix bug marking states as unreachable when defined before a Choice state
Validation for state reachability was dependent on the order that
they were defined in the document. A state declared before the Choice
Rule referencing it was marked as an unreachable state.

To simply state tracking, I am removing the map of unreachable states
in favour of using just the reachedStates map.